Rename to GraphVectorStore
