--- libs/community/langchain_community/graph_stores/__init__.py
@@ -1,3 +1,0 @@
-from langchain_community.graph_stores.cassandra import CassandraGraphStore
-
-__all__ = ["CassandraGraphStore"]
