+++ libs/community/langchain_community/graph_vectorstores/__init__.py
@@ -1,0 +1,3 @@
+from langchain_community.graph_vectorstores.cassandra import CassandraGraphVectorStore
+
+__all__ = ["CassandraGraphVectorStore"]
