+++ libs/core/langchain_core/graph_vectorstores/base.py
@@ -18,7 +18,7 @@ from langchain_core.callbacks import (
     CallbackManagerForRetrieverRun,
 )
 from langchain_core.documents import Document
-from langchain_core.graph_stores.links import METADATA_LINKS_KEY, Link
+from langchain_core.graph_vectorstores.links import METADATA_LINKS_KEY, Link
 from langchain_core.load import Serializable
 from langchain_core.pydantic_v1 import Field
 from langchain_core.runnables import run_in_executor
@@ -36,10 +36,10 @@ METADATA_CONTENT_ID_KEY = "content_id"
 
 
 class Node(Serializable):
-    """Node in the GraphStore."""
+    """Node in the GraphVectorStore."""
 
     id: Optional[str] = None
-    """Unique ID for the node. Will be generated by the GraphStore if not set."""
+    """Unique ID for the node. Will be generated by the GraphVectorStore if not set."""
     text: str
     """Text contained by the node."""
     metadata: dict = Field(default_factory=dict)
@@ -121,7 +121,7 @@ def nodes_to_documents(nodes: Iterable[Node]) -> Iterator[Document]:
         )
 
 
-class GraphStore(VectorStore):
+class GraphVectorStore(VectorStore):
     """A hybrid vector-and-graph graph store.
 
     Document chunks support vector-similarity search as well as edges linking
@@ -173,7 +173,7 @@ class GraphStore(VectorStore):
             metadatas: Optional list of metadatas associated with the texts.
                 If present, the metadata key `content_id` will be used as the node id.
                 The metadata key `links` shall be an iterable of
-                :py:class:`~langchain_core.graph_stores.links.Link`.
+                :py:class:`~langchain_core.graph_vectorstores.links.Link`.
             **kwargs: vectorstore specific parameters.
 
         Returns:
@@ -197,7 +197,7 @@ class GraphStore(VectorStore):
             metadatas: Optional list of metadatas associated with the texts.
                 If present, the metadata key `content_id` will be used as the node id.
                 The metadata key `links` shall be an iterable of
-                :py:class:`~langchain_core.graph_stores.links.Link`.
+                :py:class:`~langchain_core.graph_vectorstores.links.Link`.
             **kwargs: vectorstore specific parameters.
 
         Returns:
@@ -220,7 +220,7 @@ class GraphStore(VectorStore):
                 If present, the document's metadata key `content_id` will be used as
                 the node id.
                 The document's metadata key `links` shall be an iterable of
-                :py:class:`~langchain_core.graph_stores.links.Link`.
+                :py:class:`~langchain_core.graph_vectorstores.links.Link`.
 
         Returns:
             List of IDs of the added texts.
@@ -242,7 +242,7 @@ class GraphStore(VectorStore):
                 If present, the document's metadata key `content_id` will be used as
                 the node id.
                 The document's metadata key `links` shall be an iterable of
-                :py:class:`~langchain_core.graph_stores.links.Link`.
+                :py:class:`~langchain_core.graph_vectorstores.links.Link`.
 
         Returns:
             List of IDs of the added texts.
@@ -471,8 +471,8 @@ class GraphStore(VectorStore):
                 "'mmr' or 'traversal'."
             )
 
-    def as_retriever(self, **kwargs: Any) -> "GraphStoreRetriever":
-        """Return GraphStoreRetriever initialized from this GraphStore.
+    def as_retriever(self, **kwargs: Any) -> "GraphVectorStoreRetriever":
+        """Return GraphVectorStoreRetriever initialized from this GraphVectorStore.
 
         Args:
             search_type (Optional[str]): Defines the type of search that
@@ -489,7 +489,7 @@ class GraphStore(VectorStore):
                     lambda_mult: Diversity of results returned by MMR;
                         1 for minimum diversity and 0 for maximum. (Default: 0.5)
         Returns:
-            Retriever for this GraphStore.
+            Retriever for this GraphVectorStore.
 
         Examples:
 
@@ -526,14 +526,14 @@ class GraphStore(VectorStore):
             docsearch.as_retriever(search_kwargs={'k': 1})
 
         """
-        return GraphStoreRetriever(vectorstore=self, **kwargs)
+        return GraphVectorStoreRetriever(vectorstore=self, **kwargs)
 
 
-class GraphStoreRetriever(VectorStoreRetriever):
-    """Retriever class for GraphStore."""
+class GraphVectorStoreRetriever(VectorStoreRetriever):
+    """Retriever class for GraphVectorStore."""
 
-    vectorstore: GraphStore
-    """GraphStore to use for retrieval."""
+    vectorstore: GraphVectorStore
+    """GraphVectorStore to use for retrieval."""
     search_type: str = "traversal"
     """Type of search to perform. Defaults to "traversal"."""
     allowed_search_types: ClassVar[Collection[str]] = (
